Added XIVGear Importing

+++ BiSTracker/Models/MeldedMateria.cs
@@ -1,19 +1,15 @@
-using Lumina.Excel;
 using Lumina.Excel.GeneratedSheets;
 
 namespace BiSTracker.Models;
 
 public class MeldedMateria{
-	public MeldedMateria(ushort materiaID){
+	public MeldedMateria(uint materiaID){
         this.materiaID = materiaID;
 		hasMateria = false;
     }
 	
-	public ushort materiaID;
+	public uint materiaID;
 	public bool hasMateria;
-
-	// public ushort tempMateriaID;
-	// public byte grade;
 
 	public Materia? Row() {
 		if (!Data.CheckSheets())
